Extract generic text utilities
